Attempt to fix truncated Jedi LSP responses
This patch attempts to fix the truncated Jedit LSP responses which
produce warning messages about content-lenght mismatch. It does so
by looping around and trying again upon an empty read. The theory
is that the content producer is just not ready with more data yet.

- add function parameter 'max_empty_parts_in_a_row' with a default
value of 50.  This will be decremented on each empty read.
- change the default value of the 'max_parts' parameter to 100
since we should get larger chunks on each successful read.

Either of these values could be adjusted based on experience.