Delay creation of `.virtual_documents` until actually needed
because deployments may primarily use a well-behaved
server that does not need virtual documents folder
together with secondary servers that do require it
but are only used from time to time when opening
specific file types.

This can also marginally improve startup times as there is no
need to wait on file system operation of purging the directory
which is now only done on first relevant LSP message.