Robustify handling of special cases in completion provider
Notably, we no longer adjust start/end position based on individual
items because that was not predictable; instead we revert the logic
and populate insert text as needed to match start/end. We only change
end to equal start if there are absolutely no prefix matches, which
is required because otherwise everything would be filtered out.